+++ src/pybdsim/__init__.py
@@ -52,11 +52,11 @@ Copyright Royal Holloway, University of London 2023.
 """
 
 try:
-    from ._version import version as __version__
-    from ._version import version_tuple
+    from ._version import __version__
+    from ._version import __version_tuple__
 except ImportError:
     __version__ = "unknown version"
-    version_tuple = (0, 0, "unknown version")
+    __version_tuple__ = (0, 0, "unknown version")
 
 from . import Beam
 from . import Builder